@@ -11,10 +11,10 @@ use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\ArrayExpression;
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
-use Yiisoft\Db\Expression\ExpressionBuilderTrait;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Expression\JsonExpression;
 use Yiisoft\Db\Query\Query;
+use Yiisoft\Db\Query\QueryBuilderInterface;
 use Yiisoft\Db\Query\QueryInterface;
 use Yiisoft\Json\Json;
 
@@ -23,7 +23,9 @@ use Yiisoft\Json\Json;
  */
 final class JsonExpressionBuilder implements ExpressionBuilderInterface
 {
-    use ExpressionBuilderTrait;
+    public function __construct(private QueryBuilderInterface $queryBuilder)
+    {
+    }
 
     /**
      * Method builds the raw SQL from the $expression that will not be additionally escaped or quoted.
@@ -68,6 +70,6 @@ final class JsonExpressionBuilder implements ExpressionBuilderInterface
             return '';
         }
 
-        return '::' . (string) $expression->getType();
+        return '::' . (string)$expression->getType();
     }
 }
